Include a new resource ala XenServers for XCP1.6
Since the new system VMs use debian 6.0 it's hard to get the old
XcpServerResource to behave right for Xcp 1.6. Hypervisor Version
information is not maintained by the resource making it difficult to
extend as guest OS types change and our systemVMs progress.

The fix will allow XCP 1.6 to work with the new systemVM templates that
use Debian Squeeze and the older XCPs to work as is.

Signed-off-by: Prasanna Santhanam <tsp@apache.org>